feat: modal dialog init
