+++ package.json
@@ -62,7 +62,7 @@
     "taze": "^0.8.5",
     "typescript": "^4.9.4",
     "unocss": "^0.48.3",
-    "unplugin-vue-macros": "^1.0.3",
+    "unplugin-vue-macros": "^1.3.2",
     "vite": "^4.0.4",
     "vitest": "^0.27.0",
     "vue-tsc": "^1.0.24"
