+++ src/components/search/Search.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+import { ref } from 'vue'
+import ModalDialog from '@/components/dialog/ModalDialog.vue'
+
+const isOpenSearch = ref(true)
+</script>
+
+<template>
+  <ModalDialog v-model="isOpenSearch" />
+</template>
+
+<style scoped></style>
